Fix C++11/Clang 3.3 violation

+++ src/modelloader/vertexbufferfactory.cpp
@@ -61,7 +61,7 @@ const VertexBuffer& VertexBufferFactory::operator[](std::string objName) {
 	glBindBuffer(GL_ARRAY_BUFFER, bos[2]);
 	glBufferData(GL_ARRAY_BUFFER, vertexNormalsData.size() * sizeof(glm::vec3), &(vertexNormalsData.front()), GL_STATIC_DRAW);
 
-	vbopool.emplace(std::make_pair(std::string(objName), std::move(VertexBuffer{bos[0], bos[1], bos[2], indicesData.size() * 3, extremes})));
+	vbopool.emplace(std::make_pair(std::string(objName), std::move(VertexBuffer{bos[0], bos[1], bos[2], static_cast<unsigned int>(indicesData.size() * 3), extremes})));
 
 	/*auto it = config::preservationRules.find(objName);
 	if(it != config::preservationRules.cend()) {
